requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
google-cloud-bigquery==3.13.0
google-cloud-storage==2.14.0
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
requests==2.31.0
pyarrow==18.0.0

